[Bugfix] Disable SingleEnvThreadVerifier
During TensorIR scheduling, the `IterVar`s that represent environment
threads may duplicate, i.e. it is legal to have two env threads with
the same name tag, which may fail the `SingleEnvThreadVerifier` check
during schedule creation. This PR disables this check in this case.
In the future, it may be worthwhile to bring it back against
post-scheduling TIR.